SFDM-3384

Closes SFDM-3673

See merge request products/eafdmmart/cx-backend!86

+++ src/main/java/ru/beeline/cxbackend/domain/bi/BIInCJStep.java
@@ -7,6 +7,7 @@ package ru.beeline.cxbackend.domain.bi;
 import lombok.*;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Builder
 @Getter
@@ -31,7 +32,7 @@ public class BIInCJStep {
     private Long cjStepId;
 
     @Column(name = "\"order\"")
-    private Long order;
+    private BigDecimal order;
 
     @Column(name = "id_bi", insertable = false, updatable = false)
     private Long biId;
