+++ src/main/java/ru/beeline/cxbackend/repository/CJStepRepository.java
@@ -11,6 +11,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ru.beeline.cxbackend.domain.cj.CJStep;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
@@ -20,7 +21,7 @@ public interface CJStepRepository extends JpaRepository<CJStep, Long> {
 
     List<CJStep> findAllByCjIdIn(List<Long> cjIds);
 
-    CJStep findByCjIdAndOrder(Long cjId, Integer order);
+    CJStep findByCjIdAndOrder(Long cjId, BigDecimal order);
 
     void deleteAllByCjId(Long cjId);
